Fix for executable comment bug MDEV-8835.
When a new table is created, during CREATE TABLE and some ALTER TABLE
operations, the table definition is written to a dict entry.  If the
table has foreign key contraints, then the CONSTRAINT clauses from the
CREATE TABLE statement are used to construct the constraint definitions.
In order to do this, the comments are first stripped from the statement.
Since executable comments look like comments, the entire statement was
being stripped.  This fix removes the executable comment delimiters from
a copy of the statement before stripping comments.